# fix: optimize zero constant detection and assignment