# Description

Found in fuzzing test:
```go
type C1_6314_8435483601975321392 struct {
	X0V frontend.Variable
	Y0V frontend.Variable `gnark:",public"`
}

func (circuit *C1_6314_8435483601975321392) Define(api frontend.API) error {
	x1 := api.IsZero(circuit.X0V) // depends on X0V
	x2 := api.And(0, x1)          // always 0
	x3 := api.Cmp(x2, 1)          // always -1
	api.AssertIsEqual(1, x3)      // always false
	return nil
}

type C2_6314_8435483601975321392 struct {
	X0V frontend.Variable
	Y0V frontend.Variable `gnark:",public"`
}

func (circuit *C2_6314_8435483601975321392) Define(api frontend.API) error {
	x1 := api.Cmp(circuit.X0V, circuit.X0V) // always 0
	x2 := api.Cmp(x1, -1)                   // always -1
	x3 := api.IsZero(x2)                    // always 0
	x4 := api.IsZero(circuit.X0V)           // depends on X0V
	x5 := api.IsZero(circuit.X0V)           // depends on X0v
	x6 := api.Select(x3, x4, x5)            // always returns x5
	x7 := api.And(0, x6)                    // always 0
	api.AssertIsLessOrEqual(1, x7)          // always fails
	return nil
}
```
Where previously R1CS compiler didn't fail for second circuit but did for SCS.

## Type of change

<!-- Please delete options that are not relevant. -->

- [x] Bug fix (non-breaking change which fixes an issue)

# How has this been benchmarked?

Ran BW6 in BN254 Groth16 verifier circuit to benchmark compile time (as we now do big loop for constant value checking).
```
              │  old.txt   │           new.txt           │
              │   sec/op   │   sec/op    vs base         │
BW6InBN254-16   9.434 ± 2%   9.501 ± 1%  ~ (p=0.589 n=6)

              │   old.txt    │            new.txt            │
              │     B/op     │     B/op      vs base         │
BW6InBN254-16   21.28Gi ± 0%   21.28Gi ± 0%  ~ (p=1.000 n=6)

              │   old.txt   │           new.txt            │
              │  allocs/op  │  allocs/op   vs base         │
BW6InBN254-16   91.48M ± 0%   91.48M ± 0%  ~ (p=0.937 n=6)
```

# Checklist:

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I did not modify files generated from templates
- [x] `golangci-lint` does not output errors locally
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules